engine/requirements.txt
black==25.1.0
fastjsonschema==2.21.1
flake8==7.2.0
isort==6.0.1
pillow==11.2.1
pre-commit==4.2.0